Refactor liberime, #33
